Used Tailwind in Logo Component and deleted the css file
