--- src/components/Logo/index.css
@@ -1,8 +1,0 @@
-.logo-container {
-    width: 65%;
-    margin: 0 auto;
-}
-
-.logo {
-    width: 100%;
-}
